requirements.txt
discord.py
python-dotenv
mcstatus
proxmoxer
requests
